backend/requirements.txt
fastapi==0.100.0
uvicorn==0.23.2
python-multipart==0.0.6
pydantic==1.10.13
python-dotenv==1.0.0
openai==1.70.0
PyMuPDF==1.23.8  # This is the fitz package
python-jose==3.3.0